Update README.md
My project is based on simple library management. This program majorly focuses on managing a library or a collection of books from a manager's point of view. Using this program a manager will be able to manage and record the database of the library  and the collection of books. The manager can also add, edit or retrieve data based on the choices given in the program.